tests/wpt/tests/html/canvas/offscreen/text/2d.text.measure.width.nullCharacter.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.measure.width.nullCharacter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.measure.width.nullCharacter</h1>
<p class="desc">Null character does not take up space</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var f = new FontFace("CanvasTest", "url('/fonts/CanvasTest.ttf')");
  f.load();
  document.fonts.add(f);
  await document.fonts.ready;
  ctx.font = '50px CanvasTest';
  _assertSame(ctx.measureText('\u0000').width, 0, "ctx.measureText('\\u0000').width", "0");

}, "Null character does not take up space");
</script>
